+++ tests/math/linalg_test.py
@@ -9,7 +9,7 @@ def test_intertwiners():
     A = torch.randn(9, 9)
     def rep(q):
         return A @ G.rep(4)(q) @ A.T
-    B = intertwiners(G, G.rep(4), rep)
+    B = intertwiners(G, G.rep(4), rep, dtype=torch.get_default_dtype())
     assert torch.allclose(A, B)